server/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
pymysql>=1.1.0
bcrypt>=4.1.0
PyJWT>=2.8.0
groq>=0.4.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
